+++ ann/backpropagation.cuh
@@ -11,8 +11,8 @@
 #define NNLIB_BACKPROPAGATION_CUH
 
 void computeGradients(Layer& layer, const Matrix& delta, const Matrix& previousWeights,
-                     int batchSize, bool isLastLayer);
+                      size_t batchSize, bool isLastLayer);
 
-void applyGradients(Layer& layer, int batchSize, DTYPE learningRate = 0.01);
+void applyGradients(Layer& layer, size_t batchSize, DTYPE learningRate = 0.01);
 
 #endif //NNLIB_BACKPROPAGATION_CUH
